Rename FIR checkers package to diagnostics
